+++ src/icrar/leap-accelerate/common/constants.h
@@ -20,26 +20,7 @@
  * MA 02111 - 1307  USA
  */
 
-#include "utils.h"
-
-#include <icrar/leap-accelerate/model/MetaData.h>
-
-using namespace casacore;
-
 namespace icrar
 {
-    std::unique_ptr<MeasurementSet> ParseMeasurementSet(std::istream& input)
-    {
-        // don't skip the whitespace while reading
-        std::cin >> std::noskipws;
-
-        // use stream iterators to copy the stream to a string
-        std::istream_iterator<char> it(std::cin);
-        std::istream_iterator<char> end;
-        std::string results = std::string(it, end);
-
-        std::cout << results;
-
-        return std::make_unique<MeasurementSet>(results);
-    }
+    constexpr double speed_of_light = 299792458.0;
 }
