+++ src/icrar/leap-accelerate/tests/test_helper.h
@@ -22,7 +22,7 @@
 
 #pragma once
 
-#include <icrar/leap-accelerate/cuda/MetaDataCuda.h>
+#include <icrar/leap-accelerate/model/cuda/MetaDataCuda.h>
 
 #include <eigen3/Eigen/Core>
 #include <eigen3/Eigen/Dense>
